# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API Specification
  decription: |
    Keep in touch with your friends by sharing photos of special moments,
    thanks to WASAPhoto! You can upload your photos directly from your PC,
    and they will be visible to everyone following you.
  version: "1.0.0"
servers:
  - url: "http://localhost:3000"
    decription: "Local hosting project"
...
paths:
...
  /users/{uid}/follows:
    description: List of a user follows.
    parameters:
      - $ref: "#/components/parameters/uid"
    get:
      tags: ["follow"]
      operationId: getFollows
      summary: List user's follows.
      description: |
        Retrieve a list of users followed by this user.
      responses:
        "200": {$ref: "#/components/responses/FollowListResult"}
        "400": {$ref: "#/components/responses/GenericBadRequest"}
        "401": {$ref: "#/components/responses/BadAuthentication"}
        "403": {$ref: "#/components/responses/BadAuthorization"}
        "500": {$ref: "#/components/responses/GenericServerError"}
  /users/{uid}/follower:
    description: List of a user followers.
    parameters:
      - $ref: "#/components/parameters/uid"
    get:
      tags: ["follow"]
      operationId: getFollowers
      summary: List user's followers.
      description: |
        Retrieve a list of users following this user.
      responses:
        "200": {$ref: "#/components/responses/FollowListResult"}
        "400": {$ref: "#/components/responses/GenericBadRequest"}
        "401": {$ref: "#/components/responses/BadAuthentication"}
        "403": {$ref: "#/components/responses/BadAuthorization"}
        "500": {$ref: "#/components/responses/GenericServerError"}
  /users/{uid}/follower/{fid}:
    description: A user followed/to follow by uid.
    parameters:
      - $ref: "#/components/parameters/uid"
      - $ref: "#/components/parameters/fid"
    put:
      tags: ["follow"]
      operationId: followUser
      summary: Follow a user.
      description: |
        Follow a user -> adds it to the collection.
        As a side effect, the user followed will have this user in the follower
        section.
      responses:
        "200":
          description: User already followed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FollowData"}
        "201":
          description: User followed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FollowData"}
        "400": {$ref: "#/components/responses/GenericBadRequest"}
        "401": {$ref: "#/components/responses/BadAuthentication"}
        "403": {$ref: "#/components/responses/BadAuthorization"}
        "404":
          description: The user specified with fid does not exists.
          content:
            text/plain:
              schema: {$ref: "#/components/schemas/GenericError"}
        "500": {$ref: "#/components/responses/GenericServerError"}
    delete:
      tags: ["follow"]
      operationId: unfollowUser
      summary: Unfollow a user.
      description: |
        Unfollow a user -> removes it to the collection.
        As a side effect, the user unfollowed will have this user removed
        from his follower collection.
      responses:
        "204":
          description: User unfollowed.
        "400": {$ref: "#/components/responses/GenericBadRequest"}
        "401": {$ref: "#/components/responses/BadAuthentication"}
        "403": {$ref: "#/components/responses/BadAuthorization"}
        "500": {$ref: "#/components/responses/GenericServerError"}
# Bans logic
...
components:
  parameters:
    uid:
      name: uid
      required: true
      in: path
      description: The unique ID of a user.
      schema: {$ref: "#/components/schemas/ResourceId"}
    fid:
      name: fid
      required: true
      in: path
      description: The unique ID of a followed/following user. 
      schema: {$ref: "#/components/schemas/ResourceId"}
...
  responses:
    GenericBadRequest:
      description: |
        The request is not consistent with API specifications.
      content:
        text/plain:
          schema: {$ref: "#/components/schemas/GenericError"}
    GenericServerError:
      description: |
        Unexpected generic error on the server.
      content:
        text/plain:
          schema: {$ref: "#/components/schemas/GenericError"}
    BadAuthentication:
      description: |
          The request needs a valid authentication,
          but none (or an invalid one) was given.
      content:
        text/plain:
          schema: {$ref: "#/components/schemas/GenericError"}
    BadAuthorization:
      description: |
        The authentication provided is not sufficient to
        provide this action.
      content:
        text/plain:
          schema: {$ref: "#/components/schemas/GenericError"}
...
    FollowListResult:
      description: |
        A list of user follow.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/UserSimpleDataList"}
...
  schemas:
...
    GenericError:
      description: Generic error description.
      type: string
      minLength: 0
      maxLength: 250
      readOnly: true
    ResourceId:
      description: A generic resource 64-bit ID as a base-16 string.
      type: string
      example: "517dc7caa704a06f"
      minLength: 1
      maxLength: 16
      readOnly: true
      pattern: ^[0-9a-fA-F]{1,16}
    Username:
      description: User's personal and unique username.
      type: string
      example: john_doe_42
      pattern: ^[a-z_0-9]{3,16}
      minLength: 3
      maxLength: 16
...
    FollowData:
      description: |
        Ids representing user followed and user following.
        followingId := user following
        followedId := user followed
      type: object
      properties:
        followingId: {$ref: "#/components/schemas/ResourceId"}
        followedId: {$ref: "#/components/schemas/ResourceId"}
...
    UserSimpleDataList:
      description: |
        A list of simple user's data.
      allOf:
        - description: Retrieved users info.
          type: object
          readOnly: true
          properties:
            data:
              description: List of users data.
              type: array
              minItems: 0
              maxItems: 20
              items: {$ref: "#/components/schemas/UserSimpleData"}
...
    UserSimpleData:
      description: |
        A simple rapresentation of a user.
      type: object
      properties:
        uid: {$ref: "#/components/schemas/ResourceId"}
        username: {$ref: "#/components/schemas/Username"}
...
  securitySchemes:
    BearerToken:
      description: |
        User token used for authentication.
        Hex string base-64 created on login.
        Unsecure, used only for demostration purpose.
      type: http
      scheme: bearer
        
# Global security scheme, disabled when not necessary.
security:
  - BearerToken: []